refactor new nav
