+++ frontend/src/components/nav/nav-desktop.jsx
@@ -1,9 +1,10 @@
 import React from 'react'
 import { Col, Row } from 'antd'
-import { UIStore } from '../store'
-import { CloseIcon } from '../components/icons'
+import { UIStore } from '../../store'
+import { CloseIcon } from '../icons'
 import { motion, AnimatePresence } from 'framer-motion'
 import Link from 'next/link'
+import Item from './item'
 
 const navVariants = {
   open: { scale: 1, opacity: 1 },
@@ -30,80 +31,6 @@ const menuItemVariants = {
   },
 }
 
-export const FullscreenNav = ({ isOpen, toggle, contentKey }) => {
-  let ContentComponent
-
-  switch (contentKey) {
-    case 'Tools':
-      ContentComponent = ToolsMenu
-      break
-    default:
-      ContentComponent = () => <div>Select a menu item...</div>
-  }
-
-  return (
-    <AnimatePresence>
-      {isOpen && (
-        <motion.div
-          initial="closed"
-          animate="open"
-          exit="closed"
-          variants={navVariants}
-          transition={{ duration: 0.5 }}
-          className="fullscreen-nav"
-        >
-          <button onClick={toggle}>
-            <CloseIcon />
-          </button>
-          <ContentComponent />
-        </motion.div>
-      )}
-    </AnimatePresence>
-  )
-}
-
-export const Item = ({
-  title,
-  subtitle,
-  icon,
-  iconClass,
-  to,
-  href,
-  setShowMenu,
-}) => {
-  const contents = (
-    <>
-      <div className={['icon', iconClass].filter((it) => it != null).join(' ')}>
-        {icon}
-      </div>
-      <div className="content">
-        <b className="p-s">{title}</b>
-        <span>{subtitle}</span>
-      </div>
-    </>
-  )
-
-  if (to != null) {
-    return (
-      <Link href={to} legacyBehavior>
-        <a>{contents}</a>
-      </Link>
-    )
-  } else if (href != null) {
-    return <a href={href}>{contents}</a>
-  }
-
-  return (
-    <>
-      <div className="icon">{icon}</div>
-      <div className="content">
-        <b className="p-s">{title}</b>
-        <span>{subtitle}</span>
-      </div>
-    </>
-  )
-}
-
 const ToolsMenu = () => {
   const { menuList } = UIStore.useState((s) => ({
     menuList: s.menuList,
@@ -112,8 +39,8 @@ const ToolsMenu = () => {
     <div className="container sub-menu">
       <Row gutter={[168, 168]}>
         {menuList
-          .find((item) => item.key === 'Tools')
-          ?.children.map((menu) => (
+          .find((it) => it.key === 'Tools')
+          .children.map((menu) => (
             <Col span={8} key={menu.key}>
               <motion.p
                 className="p-m"
@@ -146,3 +73,37 @@ const ToolsMenu = () => {
     </div>
   )
 }
+
+const NavDesktop = ({ isOpen, toggle, contentKey }) => {
+  let ContentComponent
+
+  switch (contentKey) {
+    case 'Tools':
+      ContentComponent = ToolsMenu
+      break
+    default:
+      ContentComponent = () => <div>Select a menu item...</div>
+  }
+
+  return (
+    <AnimatePresence>
+      {isOpen && (
+        <motion.div
+          initial="closed"
+          animate="open"
+          exit="closed"
+          variants={navVariants}
+          transition={{ duration: 0.5 }}
+          className="fullscreen-nav"
+        >
+          <button onClick={toggle}>
+            <CloseIcon />
+          </button>
+          <ContentComponent />
+        </motion.div>
+      )}
+    </AnimatePresence>
+  )
+}
+
+export default NavDesktop
