+++ frontend/src/store.js
@@ -25,17 +25,17 @@ const geoCoverageTypeOptions = [
 ]
 
 const menuList = [
-  {
-    key: 'Plastic',
-    children: [
-      {
-        key: 'Topics',
-      },
-      {
-        key: 'Basics',
-      },
-    ],
-  },
+  // {
+  //   key: 'Plastic',
+  //   children: [
+  //     {
+  //       key: 'Topics',
+  //     },
+  //     {
+  //       key: 'Basics',
+  //     },
+  //   ],
+  // },
   {
     key: 'Tools',
     children: [
@@ -145,14 +145,14 @@ const menuList = [
       },
     ],
   },
-  {
-    key: 'Countries',
-    children: [
-      {
-        key: 'Information',
-      },
-    ],
-  },
+  // {
+  //   key: 'Countries',
+  //   children: [
+  //     {
+  //       key: 'Information',
+  //     },
+  //   ],
+  // },
   {
     key: 'About Us',
     children: [
